Reorder commands

+++ resources/migrations/03_stock.up.sql
@@ -5,5 +5,5 @@ CREATE TABLE stock  (
     exchange_id UUID REFERENCES exchange(id),
     name VARCHAR(120) NOT NULL,
     symbol VARCHAR(10) NOT NULL,
-    value FLOAT DEFAULT 0
+    value NUMERIC(7, 3) DEFAULT 0
 );
