+++ resources/migrations/05_stock_dividend_yield_column.up.sql
@@ -1,2 +1,2 @@
 -- Add new column.
-ALTER TABLE stock ADD dividend_yield FLOAT DEFAULT 0;
+ALTER TABLE stock ADD dividend_yield NUMERIC(7, 2) DEFAULT 0;
